Added a lot
Login works,
added files such as generic_fns which is a collective of other files in order to make it easy to include in other big files/pages
index makes u go to the login screen when going to the website.
db_fns which provides the function to connect to the database per file
Edited the login verification query to fit what the current database i have is.
I input the user.sql which is the database i currently have working
made the main page which is basically just a banner used and imported from the function file i made, output_fns which is a file to just output pieces of html code